Customer class written
